fix: relocate ParameterHelper class.
