+++ tests/data/tasks/xlsx/output/catalog.json
@@ -4,7 +4,7 @@
     "metadata": {
       "title": "Component Parameters",
       "last-modified": "2021-07-19T14:03:03.000+00:00",
-      "version": "0.20.0",
+      "version": "0.21.0",
       "oscal-version": "1.0.0"
     },
     "params": [
