+++ trestle/tasks/xlsx_to_oscal_component_definition.py
@@ -29,7 +29,12 @@ from openpyxl.worksheet.worksheet import Worksheet
 
 from trestle import __version__
 from trestle.oscal import OSCAL_VERSION
+from trestle.oscal.catalog import Catalog
+from trestle.oscal.common import Link
 from trestle.oscal.common import Metadata
+from trestle.oscal.common import Parameter
+from trestle.oscal.common import ParameterGuideline
+from trestle.oscal.common import ParameterValue
 from trestle.oscal.common import Party
 from trestle.oscal.common import Property
 from trestle.oscal.common import Remarks
@@ -45,7 +50,6 @@ from trestle.oscal.component import Statement
 from trestle.tasks.base_task import TaskBase
 from trestle.tasks.base_task import TaskOutcome
 from trestle.utils.oscal_helper import CatalogHelper
-from trestle.utils.parameter_helper import ParameterHelper
 
 logger = logging.getLogger(__name__)
 
@@ -807,3 +811,69 @@ class XlsxToOscalComponentDefinition(TaskBase):
         if value is None:
             raise RuntimeError(f'row {row} col {col} missing component name')
         return value
+
+
+t_guidelines = str
+t_href = str
+t_id = str
+t_label = str
+t_ofile = str
+t_oscal_version = str
+t_parameters = Dict[str, Parameter]
+t_timestamp = str
+t_usage = str
+t_values = Any
+t_verbose = bool
+t_version = str
+
+
+class ParameterHelper():
+    """Parameter Helper class is a temporary hack because Component Definition does not support Parameters."""
+
+    def __init__(
+        self, values: t_values, id_: t_id, label: t_label, href: t_href, usage: t_usage, guidelines: t_guidelines
+    ) -> None:
+        """Initialize."""
+        self._parameter_values = ParameterValue(__root__=str(values))
+        self._id = id_
+        self._label = label
+        self._links = [Link(href=href)]
+        self._usage = usage
+        self._guidelines = ParameterGuideline(prose=guidelines)
+
+    def get_parameter(self) -> Parameter:
+        """Get parameter."""
+        parameter = Parameter(
+            id=self._id,
+            label=self._label,
+            links=self._links,
+            usage=self._usage,
+            guidelines=[self._guidelines],
+            values=[self._parameter_values]
+        )
+        return parameter
+
+    def write_parameters_catalog(
+        self,
+        parameters: t_parameters,
+        timestamp: t_timestamp,
+        oscal_version: t_oscal_version,
+        version: t_version,
+        ofile: t_ofile,
+        verbose: t_verbose,
+    ) -> None:
+        """Write parameters catalog."""
+        parameter_metadata = Metadata(
+            title='Component Parameters',
+            last_modified=timestamp,
+            oscal_version=oscal_version,
+            version=version,
+        )
+        parameter_catalog = Catalog(
+            uuid=str(uuid.uuid4()),
+            metadata=parameter_metadata,
+            params=list(parameters.values()),
+        )
+        if verbose:
+            logger.info(f'output: {ofile}')
+        parameter_catalog.oscal_write(pathlib.Path(ofile))
